Guard serviceability symlink update

+++ internal/controller/assets/helper/linperf.sh
@@ -179,10 +179,14 @@ fi
 # Create serviceability link
 ############################
 if ! test -L /liberty/logs; then
-  SERVICEABILITY_FOLDER="/serviceability/$SERVICEABILITY_NAMESPACE/$HOSTNAME/logs"
-  mkdir -p $SERVICEABILITY_FOLDER
-  rm -rf /liberty/logs
-  ln -s $SERVICEABILITY_FOLDER /liberty/logs
+  if [[ ! -z "$SERVICEABILITY_NAMESPACE" ]] && [[ ! -z $HOSTNAME ]]; then
+    SERVICEABILITY_FOLDER="/serviceability/$SERVICEABILITY_NAMESPACE/$HOSTNAME/logs"
+    mkdir -p $SERVICEABILITY_FOLDER
+    rm -rf /liberty/logs
+    ln -s $SERVICEABILITY_FOLDER /liberty/logs
+  else
+    ln -s /serviceability /liberty/logs
+  fi
 fi
 
 ################################
